// frontend/src/pages/ImageGeneration/NAI/types/nai.types.ts
export interface NAIParams {
  model: string;
  prompt: string;
  negative_prompt: string;
  resolution: string;
  steps: number;
  scale: number;
  sampler: string;
  n_samples: number;
  variety_plus: boolean;
  cfg_rescale: number;
  noise_schedule: string;
  uncond_scale: number;
}

export interface NAIUserData {
  subscription: {
    tier: number;
    active: boolean;
    tierName: string;
  };
  anlasBalance: number;
}

export interface NAIGenerationResponse {
  historyIds: number[];
  count: number;
  metadata: Record<string, unknown>;
}

export interface NAIGenerationRequest extends NAIParams {
  width: number;
  height: number;
  groupId?: number;
}